backend/pyproject.toml
[tool.poetry]
name = "twofish"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.6"
django = "^3.0.5"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^4.0.0"}

[tool.poetry.dev-dependencies]
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
